requirements.txt
# Optimized requirements file

# Core dependencies
alembic==1.12.0
fastapi==0.109.2
fastapi-pagination==0.12.10
uvicorn==0.23.2
starlette==0.36.3
sqlalchemy==2.0.20
FastAPI-SQLAlchemy==0.2.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pandas==2.1.1
redis==5.0.1


# Other utilities
requests==2.31.0
PyYAML==6.0.1
reportlab==4.0.6

# Compatibility and support
annotated-types==0.5.0
anyio==3.7.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pyjwt==2.8.0
six==1.16.0
sniffio==1.3.0
sqlalchemy-spanner==1.6.2
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
watchfiles==0.20.0
sqlmodel
python-jose==3.3.0
passlib==1.7.4
